APP-4206 Add an interfaces for the Props
